Bezier movement tree
Pathing is no longer static. The Bezier curve takes over the movement. Rotation is same but more complicated. A lot of things are currently tied together for the vertical slice. Afterwards this will be updated and optimised. No bugs were encountered. Rotation points aren't perfect but they are good enough and if we receive a new site we might have to relocate anyway so until further notice this should be ok.
Important note:
- if anyone wants to rearange the pathing, contact me for instructions.