Relationship functions changed
The old system worked on an individual slider for each and every worker that was shown on screen on a scale of 0 - 100.

After feedback, it was decided relationships would be better not shown, but interpreted by the player. so we now have two sliders, Good & Bad, tracking every decision made inside the game.

Very Good = +2 GoodMeter
Good = +1 GoodMeter
Very Bad = +2 BadMeter
Bad = +1 BadMeter

This way we can use the slider values to track good and bad decisions which can be used for scripted events and/or end of game results.